+++ src/main/java/com/example/order_management_system/dto/shopping_cart/request/ShoppingCartRequest.java
@@ -17,7 +17,7 @@ import java.util.List;
 public class ShoppingCartRequest {
 
     @JsonProperty("id_owner_shopping_cart")
-    private long idOwnerShoppingCart;
+    private int idOwnerShoppingCart;
 
     @JsonProperty("item_list")
     @NotNull(message = "{application.dto.notNull}")
